todo-list: fixed the bugs, added some additional features
