--- mentorship/tasks/todo-list-enhancement/src/index.css
@@ -1,27 +1,0 @@
-body {
-  width: 1920px;
-  scroll-behavior: smooth;
-  font-family: "Roboto", sans-serif;
-  background-color: #f2f2f2;
-  overflow-x: hidden;
-}
-.main-page {
-  display: flex;
-  align-items: flex-start;
-  justify-content: flex-start;
-}
-.todo-list {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-.progress-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.todo-list.hidden > *:not(.modal-window-container) {
-  z-index: 0;
-  filter: blur(10px);
-}
