--- mentorship/tasks/todo-list-enhancement/src/styles/calendar.css
@@ -1,48 +1,0 @@
-.calendar-container {
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  flex-direction: column;
-}
-.calendar-week:hover {
-  box-shadow: 1px 1px 10px rgba(0, 0, 0, 0.3);
-}
-.calendar-week {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  gap: 20px;
-  list-style: none;
-  align-items: center;
-  text-align: center;
-  background-color: white;
-  padding-right: 200px;
-  padding-left: 200px;
-  box-sizing: border-box;
-  border-radius: 10px;
-  padding-top: 20px;
-  padding-bottom: 20px;
-  box-shadow: 1px 1px 8px rgba(0, 0, 0, 0.15);
-  transition: 0.15s ease-in-out;
-}
-.single-day {
-  font-weight: bold;
-  background-color: lightblue;
-  border-radius: 5px;
-  padding: 10px 20px 10px 20px;
-  cursor: pointer;
-  transition: 0.15s ease-in-out;
-}
-.single-day:hover {
-  box-shadow: 1px 1px 8px rgba(0, 0, 0, 0.35);
-}
-.current-day {
-  background-color: lightseagreen;
-  border-radius: 5px;
-  padding: 20px;
-  color: white;
-}
-.calendar-current-date h1 {
-  margin: 0;
-  opacity: 0.5;
-}
