--- mentorship/tasks/todo-list-enhancement/src/styles/greetings.css
@@ -1,14 +1,0 @@
-.welcome-message-container {
-  display: flex;
-  flex-direction: column;
-}
-.welcome-message-container h1,
-.welcome-message-container h2 {
-  margin: 5px 0 10px -2px;
-}
-.welcome-message-container h2 {
-  font-size: 30px;
-}
-.welcome-message-container h1 {
-  font-size: 35px;
-}
