--- mentorship/tasks/todo-list-enhancement/src/styles/list.css
@@ -1,95 +1,0 @@
-.tasks-task-list {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-.tasks-task-list ul li {
-  display: flex;
-  margin-top: 10px;
-  list-style: none;
-  background-color: white;
-  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.15);
-  align-items: center;
-  border-radius: 12px;
-  padding: 10px 15px 10px 15px;
-  width: 810px;
-  margin-right: 108px;
-}
-.completed {
-  margin-left: 70px;
-}
-.single-task-title {
-  margin: 0px 30px 0 2px;
-  inline-size: 100%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-.single-task-title:hover::before {
-  color: inherit;
-  content: attr(data-title);
-  position: fixed;
-  margin-top: 30px;
-  background-color: white;
-  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.25);
-  width: fit-content;
-  max-width: 200px;
-  overflow-wrap: break-word;
-  padding: 10px;
-  border-radius: 10px;
-}
-.single-task-title::before {
-  animation: title 0.3s ease;
-}
-.single-task-icon {
-  transform: scale(3);
-  padding-right: 3px;
-  margin-top: 1.1px;
-}
-.list-span {
-  display: flex;
-  cursor: pointer;
-  margin-top: 5px;
-  margin-left: 460px;
-  gap: 10px;
-}
-.list-span span {
-  border-radius: 15px;
-  transition: 0.2s ease-in-out;
-}
-.list-span span:nth-child(1):hover {
-  color: orange;
-}
-.list-span span:nth-child(2):hover {
-  color: red;
-}
-.task-list-clear-list {
-  border: none;
-  color: white;
-  background-color: #a52a2a;
-  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.15);
-  padding: 10px 20px 10px 20px;
-  border-radius: 8px;
-  cursor: pointer;
-  transition: 0.25s ease-in-out;
-  margin-right: 50px;
-}
-.task-list-clear-list:hover {
-  background-color: darkred;
-  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.5);
-  transform: scale(1.02);
-}
-.single-task-date {
-  display: flex;
-  justify-content: flex-end;
-  color: black;
-  width: 300px;
-  margin: 0 -175px 0 0;
-}
-.single-task-date.completed-date {
-  color: green;
-  margin: 0 -100px 0 0;
-}
-.single-task-date.out-of-deadline {
-  color: red;
-}
