--- mentorship/tasks/todo-list-enhancement/src/styles/modal-window.css
@@ -1,143 +1,0 @@
-.modal-window-container.hide-modal {
-  opacity: 0;
-  visibility: hidden;
-  z-index: 0;
-}
-.modal-window-container.show-modal {
-  opacity: 1;
-  visibility: visible;
-  z-index: 9999;
-}
-
-.modal-window-container {
-  align-items: center;
-  text-align: center;
-  flex-direction: column;
-  position: fixed;
-  top: 50%;
-  right: 50%;
-  transform: translate(50%, -50%);
-  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.5);
-  border-radius: 20px;
-  background-color: white;
-  padding: 80px;
-  transition: 0.25s ease-in-out;
-  width: 250px;
-}
-
-.modal-window-container input[type="text"] {
-  border: none;
-  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.25);
-  padding: 15px;
-  border-radius: 10px;
-  width: 200px;
-}
-.modal-window-form {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
-
-.modal-window-buttons {
-  margin-top: 40px;
-  width: 100px;
-  height: 40px;
-  border: none;
-  border-radius: 10px;
-  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.1);
-  transition: 00.25s ease-in-out;
-  cursor: pointer;
-}
-
-.error-message {
-  position: absolute;
-  bottom: 550px;
-  opacity: 0;
-  visibility: hidden;
-  color: red;
-  transition: 0.3s;
-}
-.error-message.show {
-  transform: translateY(5px);
-  visibility: visible;
-  opacity: 1;
-}
-.description-textarea {
-  resize: none;
-  width: 210px;
-  height: 115px;
-  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.25);
-  border: none;
-  border-radius: 10px;
-  padding: 10px;
-}
-.desctiption-textarea-container {
-  margin: 20px 0 20px 0;
-}
-.modal-window-buttons-container {
-  display: flex;
-  gap: 10px;
-}
-.modal-window-buttons.disabled {
-  cursor: not-allowed;
-}
-.character-count-modal-window {
-  position: absolute;
-  bottom: 230px;
-  left: 270px;
-  font-size: 12px;
-}
-.character-count-modal-window.modified {
-  left: 265px;
-}
-.fields-view-completed-tasks input[type="text"] {
-  margin-bottom: 15px;
-}
-.confirmation-modal-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-.confirmation-message {
-  color: green;
-  visibility: hidden;
-  transition: 0.3s ease-in-out;
-  position: absolute;
-  bottom: 15px;
-  left: 85px;
-}
-.confirmation-message.show-message {
-  visibility: visible;
-  transform: translateY(-10px);
-}
-.overview-title {
-  overflow: hidden;
-  text-overflow: "...`";
-  inline-size: 250px;
-  display: block;
-}
-@keyframes title {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-.task-title:hover::before {
-  color: inherit;
-  content: attr(data);
-  position: absolute;
-  top: 220px;
-  left: 330px;
-  background-color: white;
-  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.25);
-  width: fit-content;
-  max-width: 200px;
-  overflow-wrap: break-word;
-  padding: 10px;
-  border-radius: 10px;
-}
-.task-title::before {
-  animation: title 0.3s ease;
-}
