--- mentorship/tasks/todo-list-enhancement/src/styles/progress.css
@@ -1,70 +1,0 @@
-.progress-section {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  gap: 135px;
-}
-.weekly-score,
-.daily-score,
-.uncompleted-tasks {
-  display: flex;
-  background-color: white;
-  flex-direction: column;
-  padding: 30px;
-  border-radius: 15px;
-  box-shadow: 1px 1px 15px rgba(0, 0, 0, 0.15);
-  transition: 0.15s ease-in-out;
-}
-.weekly-score:hover,
-.daily-score:hover,
-.uncompleted-tasks:hover {
-  box-shadow: 1px 1px 15px rgba(0, 0, 0, 0.3);
-}
-.progress-icon {
-  margin-left: 50px;
-  margin-top: 20px;
-  transform: scale(2.5);
-  padding: 4px;
-  border-radius: 5px;
-  color: white;
-  box-shadow: 1px 1px 15px rgba(0, 0, 0, 0.2);
-}
-.progress-icon.graph {
-  background-color: tomato;
-}
-.progress-icon.stairsup {
-  background-color: cyan;
-}
-.progress-icon.todocalendar {
-  background-color: lightsalmon;
-}
-.first-line,
-.second-line {
-  flex: 1;
-}
-.progress-bar {
-  margin-top: 20px;
-  align-items: center;
-  justify-content: center;
-  border-radius: 15px;
-  display: flex;
-  width: 100%;
-  height: 15px;
-  background-color: rgb(238, 238, 238);
-  box-shadow: 1px 2px 5px rgba(0, 0, 0, 0.2);
-}
-.progress-bar.zero-score {
-  color: red;
-  text-shadow: 0px 1.5px 3px rgba(0, 0, 0, 0.3);
-}
-.progress-bar strong {
-  font-size: 12px;
-  padding-left: 4px;
-  margin: 0;
-}
-.first-paragraph {
-  display: flex;
-  justify-content: space-evenly;
-  flex-direction: row;
-}
