--- mentorship/tasks/todo-list-enhancement/src/styles/tasks.css
@@ -1,78 +1,0 @@
-.tasks-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: white;
-  margin-top: 20px;
-  flex-direction: column;
-  width: 990px;
-  border-radius: 15px;
-  box-shadow: 1px 1px 15px rgba(0, 0, 0, 0.15);
-  padding-bottom: 20px;
-  transition: 0.15s ease-in-out;
-}
-.tasks-container:hover {
-  box-shadow: 1px 1px 15px rgba(0, 0, 0, 0.3);
-}
-.tasks-container h1 {
-  font-size: 30px;
-  margin-right: 55px;
-  color: violet;
-}
-.tasks-header {
-  display: flex;
-  justify-content: center;
-  flex-direction: row;
-  align-items: center;
-  gap: 50px;
-}
-.tasks-header-container input[type="text"] {
-  border: none;
-  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.15);
-  padding: 20px;
-  width: 800px;
-  border-radius: 15px;
-}
-.tasks-header-container,
-.tasks-add-new-task {
-  flex: 1;
-}
-.tasks-add-new-task {
-  display: flex;
-  transform: scale(3);
-  cursor: pointer;
-  color: white;
-  background-color: violet;
-  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.3);
-  border-radius: 5px;
-  padding: 1.8px;
-  transition: 0.2s ease-in-out;
-}
-.tasks-add-new-task:hover {
-  background-color: darkviolet;
-  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.7);
-}
-.no-values-error {
-  margin-right: 90px;
-}
-.completed-tasks-transition-button {
-  position: absolute;
-  top: 522px;
-  left: 1288px;
-  color: whitesmoke;
-  cursor: pointer;
-  transition: 0.2s ease-in-out;
-  font-size: 20px;
-  background-color: #ff729f;
-  padding: 10px;
-  border-top-right-radius: 15px;
-  box-shadow: 1px 1px 8px rgba(0, 0, 0, 0.3);
-  z-index: 9999;
-}
-.completed-tasks-transition-button:hover {
-  box-shadow: 2px 2px 10px rgba(0, 0, 0, 0.7);
-}
-.completed-tasks-transition-button.clicked {
-  width: 155px;
-  text-align: center;
-}
