+++ mentorship/tasks/todo-list/src/components/List.js
@@ -2,16 +2,14 @@ import React from "react";
 import "../styles/list.css";
 import { AiOutlineDelete, AiOutlineEdit } from "react-icons/ai";
 import { MdOutlineTaskAlt, MdDownloadDone } from "react-icons/md";
-import { useTaskProgressContext } from "./ProgressContext";
-function List({
-  list,
-  editItem,
-  setDoneTask,
-  setDeleteID,
-  setShowConfirmationModal,
-  setConfirmationState,
-}) {
-  const { formattedDate, formattedCompletedDate } = useTaskProgressContext();
+import { useTaskProgressContext } from "../context/ProgressContext";
+import { useListContext } from "../context/ListContext";
+function List() {
+  const { editItem, setDoneTask, setDeleteID, setConfirmationState } =
+    useListContext();
+  const { list, formattedCompletedDate, setShowConfirmationModal, today } =
+    useTaskProgressContext();
+
   const handleItemButtons = (id, { type }, title, date, description) => {
     if (type === "edit") {
       editItem(id);
@@ -34,14 +32,14 @@ function List({
   return (
     <ul>
       {list.map(({ id, title, date, description }) => {
-        const deadlineDate = new Date(formattedDate);
-        const taskDate = new Date(date);
-        const isOutOfDeadline = deadlineDate > taskDate;
-
+        const deadlineDate = new Date(date).setHours(0, 0, 0, 0);
+        const isOutOfDeadline = deadlineDate < today;
         return (
           <li key={id}>
             <MdOutlineTaskAlt className="single-task-icon" />
-            <p className="single-task-title">{title}</p>
+            <p className="single-task-title" data-title={title}>
+              {title}
+            </p>
             <p
               className={`single-task-date ${
                 isOutOfDeadline && `out-of-deadline`
