+++ mentorship/tasks/todo-list/src/components/ModalWindow.js
@@ -3,27 +3,26 @@ import ReactDOM from "react-dom";
 import "../styles/modal-window.css";
 import DatePicker from "react-datepicker";
 import "react-datepicker/dist/react-datepicker.css";
+import { useListContext } from "../context/ListContext";
+import { useTaskProgressContext } from "../context/ProgressContext";
 
 const MAX_LENGTH = 180;
 
-function ModalWindow(props) {
+function ModalWindow({ handleDescription }) {
   const {
     name,
     handleSubmit,
     setName,
-    showModal,
     isEditing,
     handleModalWindow,
     error,
-    selectedDate,
-    setSelectedDate,
     description,
     errorDate,
-    handleDescription,
     viewCompletedTasks,
-    completedTasks,
     viewID,
-  } = props;
+  } = useListContext();
+  const { completedTasks, selectedDate, setSelectedDate, showModal, today } =
+    useTaskProgressContext();
 
   const characterCountClassName = description.length >= 100 ? "modified" : "";
 
@@ -37,8 +36,9 @@ function ModalWindow(props) {
     <div className={`modal-window-container ${modalWindowClassName}`}>
       {!viewCompletedTasks ? (
         <>
-          <h1>
-            {isEditing ? `Editing task '${name}' ...` : "Create a new task..."}
+          <h1 className="overview-title">
+            {isEditing ? "Editing task " : "Create a new task..."}
+            {isEditing && "`" + name + "`"}
           </h1>
           <form onSubmit={handleSubmit}>
             <div className="modal-window-form">
@@ -71,6 +71,7 @@ function ModalWindow(props) {
                 className={error ? "error" : ""}
                 placeholderText="Select deadline for your task.."
                 disabled={viewCompletedTasks}
+                minDate={today}
               />
               {!viewCompletedTasks ? (
                 <div className="modal-window-buttons-container">
@@ -114,15 +115,20 @@ function ModalWindow(props) {
               if (viewID === id) {
                 return (
                   <div key={id}>
-                    <h1>{`"${name}" overview`}</h1>
+                    <h1 className="overview-title">
+                      `{name}` <br />
+                      overview
+                    </h1>
                     <div className="modal-window-form">
-                      <span>Task name:</span>
-                      <input
-                        type="text"
-                        value={name}
-                        className={error ? "error" : ""}
-                        disabled={viewCompletedTasks}
-                      />
+                      <div className="task-title" data={name}>
+                        <span>Task name:</span>
+                        <input
+                          type="text"
+                          value={name}
+                          className={error ? "error" : ""}
+                          disabled={viewCompletedTasks}
+                        />
+                      </div>
                       <div className="desctiption-textarea-container">
                         <span>Description: </span>
                         <textarea
