+++ mentorship/tasks/todo-list/src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
 import "./index.css";
-import ProgressContext from "./components/ProgressContext";
+import ProgressContext from "./context/ProgressContext";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
