+++ mentorship/tasks/todo-list/src/sections/Calendar.js
@@ -1,26 +1,9 @@
 import React from "react";
+import { useCalendarContext } from "../context/CalendarContext";
 import "../styles/calendar.css";
 
-const daysOfWeek = ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"];
-const today = new Date();
-const startOfWeek = new Date(
-  today.getFullYear(),
-  today.getMonth(),
-  today.getDate() - today.getDay()
-);
-const days = Array.from(
-  { length: 7 },
-  (_, i) =>
-    new Date(
-      startOfWeek.getFullYear(),
-      startOfWeek.getMonth(),
-      startOfWeek.getDate() + i
-    )
-);
-const year = today.getFullYear();
-const month = today.toLocaleString("en-US", { month: "long" });
-
 const Calendar = () => {
+  const { days, year, month, daysOfWeek, today } = useCalendarContext();
   return (
     <div className="calendar-container">
       <div className="calendar-current-date">
