+++ mentorship/tasks/todo-list/src/styles/list.css
@@ -22,6 +22,24 @@
 .single-task-title {
   margin: 0px 30px 0 2px;
   inline-size: 100%;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+.single-task-title:hover::before {
+  color: inherit;
+  content: attr(data-title);
+  position: fixed;
+  margin-top: 30px;
+  background-color: white;
+  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.25);
+  width: fit-content;
+  max-width: 200px;
+  overflow-wrap: break-word;
+  padding: 10px;
+  border-radius: 10px;
+}
+.single-task-title::before {
+  animation: title 0.3s ease;
 }
 .single-task-icon {
   transform: scale(3);
