+++ mentorship/tasks/todo-list/src/styles/modal-window.css
@@ -51,7 +51,7 @@
 
 .error-message {
   position: absolute;
-  bottom: 500px;
+  bottom: 550px;
   opacity: 0;
   visibility: hidden;
   color: red;
@@ -65,7 +65,7 @@
 .description-textarea {
   resize: none;
   width: 210px;
-  height: 100px;
+  height: 115px;
   box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.25);
   border: none;
   border-radius: 10px;
@@ -109,4 +109,35 @@
 .confirmation-message.show-message {
   visibility: visible;
   transform: translateY(-10px);
+}
+.overview-title {
+  overflow: hidden;
+  text-overflow: "...`";
+  inline-size: 250px;
+  display: block;
+}
+@keyframes title {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+.task-title:hover::before {
+  color: inherit;
+  content: attr(data);
+  position: absolute;
+  top: 220px;
+  left: 330px;
+  background-color: white;
+  box-shadow: 2px 2px 20px rgba(0, 0, 0, 0.25);
+  width: fit-content;
+  max-width: 200px;
+  overflow-wrap: break-word;
+  padding: 10px;
+  border-radius: 10px;
+}
+.task-title::before {
+  animation: title 0.3s ease;
 }
